Render HTML Void Elements without a closing tag as per spec
Use a union type to represent each of the 5 kinds of elements defined in the spec. Created
a helper function to tag a element name as a String and convert it to the union type.
Use pattern matching in  on the new type to control the render behavior
of the closing tags. Left a stub for correctly handling Raw Text element kind and Escaped Raw Text
element kind but did not implement correct rendering of element contents for these kinds.